+++ examples/Qcodes_drivers/basic-driver/hello_qcodes.py
@@ -216,6 +216,12 @@ def OPX_1d_scan(simulate=False):
 if run == "1d":
     # Axis1 is the most inner loop
     opx_instrument.set_sweep_parameters("axis1", biases, "V", "Biases")
+    # Update the readout length for this experiment
+    readout_length = 600  # in ns
+    opx_instrument.update_readout_length("readout_element", "readout", readout_length)
+    # Pass the readout length (in ns) to the class to convert the demodulated/integrated data into Volts
+    # and create the setpoint Parameter for raw adc trace acquisition
+    opx_instrument.readout_pulse_length(readout_length)
     # Add the custom sequence to the OPX
     opx_instrument.qua_program = OPX_1d_scan(simulate=True)
     # Simulate program
@@ -397,3 +403,93 @@ if run == "sliced":
         do_plot=True,
         exp=experiment,
     )
+
+#####################################
+#     1D OPX SWEEP & datasaver      #
+#####################################
+# Show how to perform a 1D sweep with averaging on the OPX
+# while scanning an external parameter using the qcodes Measurement framework instead of do1d.
+from qualang_tools.loops import from_array
+from qcodes.dataset import plot_dataset
+
+frequencies = np.arange(50e6, 300e6, 0.1e6)
+n_avg = 10
+
+
+# QUA sequence
+def OPX_reflectometry(simulate=False):
+    with program() as prog:
+        f = declare(int)
+        n = declare(int)
+        I = declare(fixed)
+        Q = declare(fixed)
+        Q_st = declare_stream()
+        I_st = declare_stream()
+        with infinite_loop_():
+            if not simulate:
+                pause()
+            with for_(n, 0, n < n_avg, n + 1):
+                with for_(*from_array(f, frequencies)):
+                    update_frequency("readout_element", f)
+                    measure(
+                        "readout",
+                        "readout_element",
+                        None,
+                        demod.full("cos", I, "out1"),
+                        demod.full("sin", Q, "out1"),
+                    )
+                    wait(1000, "readout_element")
+                    save(I, I_st)
+                    save(Q, Q_st)
+
+        with stream_processing():
+            I_st.buffer(len(frequencies)).buffer(n_avg).map(
+                FUNCTIONS.average()
+            ).save_all("I")
+            Q_st.buffer(len(frequencies)).buffer(n_avg).map(
+                FUNCTIONS.average()
+            ).save_all("Q")
+    return prog
+
+
+if run == "datasaver":
+    # Axis1 is the most inner loop
+    opx_instrument.set_sweep_parameters("axis1", frequencies, "Hz", "Readout frequency")
+    # Add the custom sequence to the OPX
+    opx_instrument.qua_program = OPX_reflectometry(simulate=True)
+    # Simulate program
+    opx_instrument.sim_time(10_000)
+    opx_instrument.simulate()
+    opx_instrument.plot_simulated_wf()
+    # Execute program
+    opx_instrument.qua_program = OPX_reflectometry(simulate=False)
+    # Initialize the measurement
+    meas = qc.Measurement(exp=experiment, name="reflectometry")
+    # Register the qcodes parameter that we'll sweep
+    meas.register_parameter(VP1)
+    # Get the Result parameters from the OPX scan
+    OPX_param = opx_instrument.get_measurement_parameter()
+    # Register all the involved parameters
+    meas.register_parameter(OPX_param, setpoints=[VP1])
+    # Start the sequence (Send the QUA program to the OPX which compiles and executes it)
+    opx_instrument.run_exp()
+    # Start the qcodes measurement
+    with meas.run() as datasaver:
+        # Loop over an external parameter
+        for gate_v in np.arange(0, 0.01, 0.001):
+            # Update ethe external parameter
+            VP1(gate_v)
+            # Run the QUA sequence
+            opx_instrument.resume()
+            # Get the results from the OPX
+            data = opx_instrument.get_res()
+            # Store the results in the scodes database
+            datasaver.add_result(
+                (VP1, VP1()), (OPX_param, np.array(list(data.values())))
+            )
+        # Halt the OPX program at the end
+        opx_instrument.halt()
+        # Get the full dataset
+        dataset = datasaver.dataset
+    # Plot the dataset
+    plot_dataset(dataset)